requirements.txt
# PrimeOps Agentic OS — Dependencies

# --- Web Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# --- Database ---
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
alembic>=1.14.0
psycopg2-binary>=2.9.9

# --- Data Engine ---
pandas>=2.2.0
numpy>=1.26.0

# --- Google Cloud & ADK ---
# google-adk 1.31.1 is already installed and manages google-cloud-aiplatform transitively.
# vertexai is a namespace inside google-cloud-aiplatform, not a separate package.
google-adk>=1.0.0
google-cloud-storage>=2.18.0

# --- HTTP (used by ADK tools to call local FastAPI endpoints) ---
requests>=2.31.0

# --- Config & Validation ---
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
